fixed the solution because the original one was too slow, used LCM of each period

+++ src/main/java/hr/bp/adventofcode/day08/TaskTwo.java
@@ -13,7 +13,7 @@ public class TaskTwo {
         String input = readInputForDay("day08");
         NodeNetwork game = new NodeNetwork(input);
 
-        int sum = game.stepThroughNetworkInParallel();
+        long sum = game.stepThroughNetworkInParallel();
 
         Utils.displayResult(sum);
     }
